New functionality, better perf, better compatibility, bug fixes, and free puppies (#49)
### This merge contains (but not limited to):
- Support for syslog, if available, and macOS' system logger (os_log) as
their own independent logging destination
- Massive performance gains for both stdio and log files via
optimization/refactoring (I would estimate 30%+)
- Lots of work to get things going smoothly on Windows
  - A Visual Studio 2022 solution in `msvs`
- New macros in siransimacros.h that can be used with libsir or anything
that writes to stdio (*printf, puts, etc.): `printf(RED("Oh lawd,
something " ULINE("awful") " has happened!\n");`
- The test rig now supports command-line arguments:
  - `--list` : lists all the available tests
  - `--only `: only run the specified tests
  - `--wait` : wait for keypress before exiting
  - `--perf` : runs the performance benchmark test (normally skipped)
- Bug fixes
- Potential memory leaks squashed
- Additional errors with better error reporting
- New preprocessor defines:
- `SIR_ASSERT_ENABLED`: if defined, libsir actually uses assert(), if
not, _sir_selflog outputs a red message containing the condition that
would have asserted, along with the file/line/func.
- `SIR_NO_SYSTEM_LOGGERS`: if defined, system logger support is
completely disabled/not compiled in. The entry points still exist, but
they will return an error that the feature is disabled or unavailable.

Due to the massive amount of churn that's occurred in the last couple
months, and due to the fact that I now have to rewrite a bunch of the
Doxygen comments, I will not be tagging a new release for a week or so.

Also, shout out to @johnsonjh who is now a libsir contributor! He's
going to give a whole new meaning to cross-platform compatibility for
this library (he's already done a ton of it, it's just not merged in
yet). If there's anyone else out there, I could use the help (even
though @johnsonjh types like the wind).

---------

Signed-off-by: Jeffrey H. Johnson <trnsz@pobox.com>
Co-authored-by: Jeffrey H. Johnson <trnsz@pobox.com>