Fix warning in tests
SIRTEST_CLOCK is a CLOCK source (i.e. CLOCK_REALTIME or CLOCK_MONOTONIC),
and this constant is a long long integer and not an int on many systems.

Always casting to a long long avoids the warning, and doesn't harm
performance as this only happens when formatting an error message.

Signed-off-by: Jeffrey H. Johnson <trnsz@pobox.com>