Use -g3 for most debug builds
Use -g3 for most debug builds, except for some OS's and compilers where
it's known to be problematic or unsupported.  The user can override the
default by setting the DBGFLAG variable in the environment or as an
argument to make.

Currently, all compilers on AIX, cross-compilation to AIX, and NVIDIA
HPC SDK C (on all operating systems) will keep using `-g` and not `-g3`.

Using `-g3` saves macro definitions for GDB's macro expansion support.

Signed-off-by: Jeffrey H. Johnson <trnsz@pobox.com>