Use `-g3` for most debug builds (#183)
* Use `-g3` for most debug builds, except on some systems and compilers
where it's known to be problematic or unsupported. The user can override
the default by setting the `DBGFLAG` variable in the environment (or as
an argument to make). Currently, all compilers on AIX, cross-compiling
to AIX, and NVIDIA HPC SDK C on all operating systems will continue
using `-g` and not `-g3`.
* Using `-g3` saves macro definitions and enables GDB's macro expansion
support.

Signed-off-by: Jeffrey H. Johnson <trnsz@pobox.com>