Appease latest PVS-Studio 7.29.80616.6568
The NULL pointer test now additionally raises a V575 (null pointer is
passed) in addition to the V618 (dangerous use of function), in the
`sirtest_failnulls` function in `tests/tests.c`.  Suppressing this
warning because the test does this intentionally.

Also, add `compile_commands.events.json` to the `.gitignore` file.

Signed-off-by: Jeffrey H. Johnson <trnsz@pobox.com>